// src/types/dashboard.ts
export interface SalesCardType {
    iconPath: string;
    iconColor: string;
    title: string;
    value: string;
    change: string;
    color: string;
}

export interface Product {
    id: string;
    name: string;
    progress: number;
    sales: string;
    primaryColor: string;
    secondaryColor: string;
}